requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pinecone-client==5.0.1
pinecone-plugin-assistant==1.1.0
python-multipart==0.0.9
jinja2==3.1.3